Program on if_else and operators

+++ Neelima/Day3/Practice.py
@@ -1,4 +1,4 @@
-amount=float(input("Enter the value of amount: "))
+amount=int(input("Enter the value of amount: "))
 balance=10000
 if(amount!=0 and amount>0):
     print("Credited_amount",amount+balance)
